requirements.txt
Django==3.0.5
celery==4.3.0
redis==3.2.1
vine==1.3.0
flower==0.9.5
ipython
watchdog==0.10.2
argh>=0.24.1
PyYAML>=3.10
django-celery-email==3.0.0
django-celery-beat==2.0.0
requests==2.23.0

django-capture-on-commit-callbacks
factory-boy==2.12.0
coverage==5.1

pytest==5.3.5
pytest-django==3.8.0

django-environ==0.4.5
gunicorn==20.0.4
psycopg2-binary==2.8.5

pytz==2018.5

djangorestframework==3.12.2
djangorestframework-simplejwt
django-rest-auth==0.9.5
django-silk==4.0.0
cryptography
django-cors-headers==3.7.0
boto3==1.17.12
django-storages==1.11.1
django-bootstrap4==2.3.1
django-crispy-forms==1.10.0

drf-spectacular
uritemplate
django-redis
djoser
django-extensions
Graphviz
python-jose